fix: example CAT to use test helpers.py
