+++ examples/team_recommender/tests/helpers.py
@@ -1,0 +1,15 @@
+import json
+
+from settings import root_path
+
+
+def load_json_fixture(file_name: str) -> dict:
+    """
+    Utility function to load a JSON fixture file.
+
+    :param file_name: Name of the JSON file to load.
+    :return: Parsed JSON data as a dictionary.
+    """
+    json_path = root_path() / "tests" / "fixtures" / file_name
+    with open(json_path, "r") as file:
+        return json.load(file)
